Added delay to reproduce issue 29806.
A small delay introduced in the lambdas of TransactionAddedToMempool and TransactionRemovedFromMempool is used to mimic the delayed execution of the tasks they schedule. In an actual run, the delay would be with the thread and not the lambda, but it was easier to demonstrate the effects of a delay doing it this way.